Export function to register new graph types
